test(inference): split validation session coverage
