+++ src/lib/inference/openai-validation-session-fallback.test.ts
@@ -1,0 +1,211 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import http from "node:http";
+import { describe, expect, it, vi } from "vitest";
+import {
+  type OpenAiValidationSessionDeps,
+  probeOpenAiLikeEndpointWithValidationSession,
+} from "./openai-validation-session";
+import {
+  createOpenAiValidationTestDeps,
+  useOpenAiValidationTestServers,
+} from "./openai-validation-session.test-helpers";
+
+const listen = useOpenAiValidationTestServers();
+
+describe("OpenAI validation curl fallback", () => {
+  it("replays through curl after a terminal native failure", async () => {
+    const server = http.createServer((request, response) => {
+      request.resume();
+      response.statusCode = 401;
+      response.end('{"error":{"message":"invalid key"}}');
+    });
+    const port = await listen(server);
+    const legacyProbe: OpenAiValidationSessionDeps["legacyProbe"] = vi.fn(() => ({
+      ok: false,
+      message: "curl diagnostic",
+    }));
+    const harness = createOpenAiValidationTestDeps(legacyProbe);
+
+    const result = await probeOpenAiLikeEndpointWithValidationSession(
+      `http://provider.example.test:${port}/v1`,
+      "test-model",
+      "bad-key",
+      { skipResponsesProbe: true },
+      harness,
+    );
+
+    expect(result).toEqual({ ok: false, message: "curl diagnostic" });
+    expect(legacyProbe).toHaveBeenCalledTimes(1);
+  });
+
+  it("replays through curl once after a native connection reset", async () => {
+    const server = http.createServer((request) => {
+      request.socket.destroy();
+    });
+    const port = await listen(server);
+    const legacyProbe: OpenAiValidationSessionDeps["legacyProbe"] = vi.fn(() => ({
+      ok: false,
+      message: "curl connection diagnostic",
+    }));
+    const harness = createOpenAiValidationTestDeps(legacyProbe);
+
+    const result = await probeOpenAiLikeEndpointWithValidationSession(
+      `http://provider.example.test:${port}/v1`,
+      "test-model",
+      "test-key",
+      { skipResponsesProbe: true },
+      harness,
+    );
+
+    expect(result).toEqual({ ok: false, message: "curl connection diagnostic" });
+    expect(legacyProbe).toHaveBeenCalledTimes(1);
+  });
+
+  it("replays through curl when DNS pre-resolution exceeds its deadline", async () => {
+    const legacyProbe: OpenAiValidationSessionDeps["legacyProbe"] = vi.fn(() => ({
+      ok: false,
+      message: "curl DNS diagnostic",
+    }));
+    const lookup = vi.fn(() => new Promise<Array<{ address: string; family: number }>>(() => {}));
+    const harness = createOpenAiValidationTestDeps(legacyProbe);
+    harness.sessionOptions = { env: {}, lookup, dnsTimeoutMs: 10 };
+
+    const result = await probeOpenAiLikeEndpointWithValidationSession(
+      "https://provider.example.test/v1",
+      "test-model",
+      "test-key",
+      {},
+      harness,
+    );
+
+    expect(result).toEqual({ ok: false, message: "curl DNS diagnostic" });
+    expect(lookup).toHaveBeenCalledTimes(1);
+    expect(legacyProbe).toHaveBeenCalledTimes(1);
+  });
+
+  it("replays through curl after a connection reset during Responses streaming", async () => {
+    const handleRequest = vi
+      .fn()
+      .mockImplementationOnce((_request, response) => {
+        response.end('{"output":[{"type":"message"}]}');
+      })
+      .mockImplementationOnce((request) => {
+        request.socket.destroy();
+      });
+    const server = http.createServer((request, response) => {
+      request.resume();
+      handleRequest(request, response);
+    });
+    const port = await listen(server);
+    const legacyProbe: OpenAiValidationSessionDeps["legacyProbe"] = vi.fn(() => ({
+      ok: false,
+      message: "curl streaming diagnostic",
+    }));
+    const harness = createOpenAiValidationTestDeps(legacyProbe);
+
+    const result = await probeOpenAiLikeEndpointWithValidationSession(
+      `http://provider.example.test:${port}/v1`,
+      "test-model",
+      "test-key",
+      { probeStreaming: true },
+      harness,
+    );
+
+    expect(result).toEqual({ ok: false, message: "curl streaming diagnostic" });
+    expect(handleRequest).toHaveBeenCalledTimes(2);
+    expect(legacyProbe).toHaveBeenCalledTimes(1);
+  });
+
+  it("uses curl without DNS pre-resolution when a proxy is configured", async () => {
+    const legacyProbe: OpenAiValidationSessionDeps["legacyProbe"] = vi.fn(() => ({
+      ok: true,
+      api: "openai-completions",
+    }));
+    const lookup = vi.fn();
+    const harness = createOpenAiValidationTestDeps(legacyProbe);
+    harness.sessionOptions = {
+      env: { HTTPS_PROXY: "http://proxy.example.test:8080" },
+      lookup,
+    };
+
+    const result = await probeOpenAiLikeEndpointWithValidationSession(
+      "https://provider.example.test/v1",
+      "test-model",
+      "test-key",
+      {},
+      harness,
+    );
+
+    expect(result).toMatchObject({ ok: true, api: "openai-completions" });
+    expect(legacyProbe).toHaveBeenCalledTimes(1);
+    expect(lookup).not.toHaveBeenCalled();
+  });
+
+  it.each([
+    "CURL_CA_BUNDLE",
+    "SSL_CERT_FILE",
+    "SSL_CERT_DIR",
+  ])("uses curl without DNS pre-resolution when %s is configured", async (envName) => {
+    const legacyProbe: OpenAiValidationSessionDeps["legacyProbe"] = vi.fn(() => ({
+      ok: true,
+      api: "openai-completions",
+    }));
+    const lookup = vi.fn();
+    const harness = createOpenAiValidationTestDeps(legacyProbe);
+    harness.sessionOptions = { env: { [envName]: "/tmp/provider-tls-config" }, lookup };
+
+    const result = await probeOpenAiLikeEndpointWithValidationSession(
+      "https://provider.example.test/v1",
+      "test-model",
+      "test-key",
+      {},
+      harness,
+    );
+
+    expect(result).toMatchObject({ ok: true, api: "openai-completions" });
+    expect(legacyProbe).toHaveBeenCalledTimes(1);
+    expect(lookup).not.toHaveBeenCalled();
+  });
+
+  it("keeps preflight-pinned endpoints on curl without native DNS", async () => {
+    const legacyProbe: OpenAiValidationSessionDeps["legacyProbe"] = vi.fn(() => ({
+      ok: true,
+      api: "openai-completions",
+    }));
+    const harness = createOpenAiValidationTestDeps(legacyProbe);
+
+    const result = await probeOpenAiLikeEndpointWithValidationSession(
+      "https://provider.example.test/v1",
+      "test-model",
+      "test-key",
+      { pinnedAddresses: ["203.0.113.10"] },
+      harness,
+    );
+
+    expect(result).toMatchObject({ ok: true, api: "openai-completions" });
+    expect(legacyProbe).toHaveBeenCalledTimes(1);
+    expect(harness.sessionOptions!.lookup).not.toHaveBeenCalled();
+  });
+
+  it("keeps DeepSeek V4 Pro on its specialized legacy streaming probe", async () => {
+    const legacyProbe: OpenAiValidationSessionDeps["legacyProbe"] = vi.fn(() => ({
+      ok: true,
+      api: "openai-completions",
+    }));
+    const harness = createOpenAiValidationTestDeps(legacyProbe);
+
+    const result = await probeOpenAiLikeEndpointWithValidationSession(
+      "https://provider.example.test/v1",
+      "deepseek-ai/deepseek-v4-pro",
+      "test-key",
+      {},
+      harness,
+    );
+
+    expect(result).toMatchObject({ ok: true, api: "openai-completions" });
+    expect(legacyProbe).toHaveBeenCalledTimes(1);
+    expect(harness.sessionOptions!.lookup).not.toHaveBeenCalled();
+  });
+});
