+++ src/lib/inference/openai-validation-session.test.ts
@@ -2,63 +2,14 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import http from "node:http";
-import { afterEach, describe, expect, it, vi } from "vitest";
-import type { CurlProbeResult } from "../adapters/http/probe";
+import { describe, expect, it } from "vitest";
+import { probeOpenAiLikeEndpointWithValidationSession } from "./openai-validation-session";
 import {
-  type OpenAiValidationSessionDeps,
-  probeOpenAiLikeEndpointWithValidationSession,
-} from "./openai-validation-session";
+  createOpenAiValidationTestDeps,
+  useOpenAiValidationTestServers,
+} from "./openai-validation-session.test-helpers";
 
-const servers: http.Server[] = [];
-
-afterEach(async () => {
-  await Promise.all(
-    servers.splice(0).map(
-      (server) =>
-        new Promise<void>((resolve) => {
-          server.close(() => resolve());
-          server.closeAllConnections();
-        }),
-    ),
-  );
-});
-
-async function listen(server: http.Server): Promise<number> {
-  servers.push(server);
-  await new Promise<void>((resolve) => server.listen(0, "127.0.0.1", resolve));
-  const address = server.address();
-  expect(address).toBeTruthy();
-  expect(typeof address).toBe("object");
-  return (address as import("node:net").AddressInfo).port;
-}
-
-const legacySuccess = (): CurlProbeResult => ({
-  ok: true,
-  httpStatus: 200,
-  curlStatus: 0,
-  body: "",
-  stderr: "",
-  message: "legacy",
-});
-
-function deps(
-  legacyProbe: OpenAiValidationSessionDeps["legacyProbe"] = vi.fn(legacySuccess),
-): OpenAiValidationSessionDeps {
-  return {
-    legacyProbe,
-    hasResponsesToolCall: (body: string) => body.includes('"type":"function_call"'),
-    hasChatCompletionsToolCall: (body: string) => body.includes('"tool_calls"'),
-    hasChatCompletionsToolCallLeak: () => false,
-    getChatPayload: (model: string) => ({ model, messages: [] }),
-    getResponsesTimeoutMs: () => 1_000,
-    getChatTimeoutMs: () => 1_000,
-    sessionOptions: {
-      env: {},
-      lookup: vi.fn(async () => [{ address: "127.0.0.1", family: 4 }]),
-      allowPrivateAddressesForTesting: true,
-    },
-  };
-}
+const listen = useOpenAiValidationTestServers();
 
 describe("OpenAI validation keepalive sequence", () => {
   it("uses one connection for Responses semantic fallback and Chat success", async () => {
@@ -78,7 +29,7 @@ describe("OpenAI validation keepalive sequence", () => {
       connections += 1;
     });
     const port = await listen(server);
-    const harness = deps();
+    const harness = createOpenAiValidationTestDeps();
 
     const result = await probeOpenAiLikeEndpointWithValidationSession(
       `http://provider.example.test:${port}/v1`,
@@ -120,7 +71,7 @@ describe("OpenAI validation keepalive sequence", () => {
       connections += 1;
     });
     const port = await listen(server);
-    const harness = deps();
+    const harness = createOpenAiValidationTestDeps();
 
     const result = await probeOpenAiLikeEndpointWithValidationSession(
       `http://provider.example.test:${port}/v1`,
@@ -148,7 +99,7 @@ describe("OpenAI validation keepalive sequence", () => {
       );
     });
     const port = await listen(server);
-    const harness = deps();
+    const harness = createOpenAiValidationTestDeps();
 
     const result = await probeOpenAiLikeEndpointWithValidationSession(
       `http://provider.example.test:${port}/v1`,
@@ -160,259 +111,6 @@ describe("OpenAI validation keepalive sequence", () => {
 
     expect(result).toMatchObject({ ok: true, api: "openai-responses" });
     expect(paths).toEqual(["/v1/responses", "/v1/responses"]);
-    expect(harness.legacyProbe).not.toHaveBeenCalled();
-  });
-
-  it("replays through curl after a terminal native failure", async () => {
-    const server = http.createServer((request, response) => {
-      request.resume();
-      response.statusCode = 401;
-      response.end('{"error":{"message":"invalid key"}}');
-    });
-    const port = await listen(server);
-    const legacyProbe: OpenAiValidationSessionDeps["legacyProbe"] = vi.fn(() => ({
-      ok: false,
-      message: "curl diagnostic",
-    }));
-    const harness = deps(legacyProbe);
-
-    const result = await probeOpenAiLikeEndpointWithValidationSession(
-      `http://provider.example.test:${port}/v1`,
-      "test-model",
-      "bad-key",
-      { skipResponsesProbe: true },
-      harness,
-    );
-
-    expect(result).toEqual({ ok: false, message: "curl diagnostic" });
-    expect(legacyProbe).toHaveBeenCalledTimes(1);
-  });
-
-  it("replays through curl once after a native connection reset", async () => {
-    const server = http.createServer((request) => {
-      request.socket.destroy();
-    });
-    const port = await listen(server);
-    const legacyProbe: OpenAiValidationSessionDeps["legacyProbe"] = vi.fn(() => ({
-      ok: false,
-      message: "curl connection diagnostic",
-    }));
-    const harness = deps(legacyProbe);
-
-    const result = await probeOpenAiLikeEndpointWithValidationSession(
-      `http://provider.example.test:${port}/v1`,
-      "test-model",
-      "test-key",
-      { skipResponsesProbe: true },
-      harness,
-    );
-
-    expect(result).toEqual({ ok: false, message: "curl connection diagnostic" });
-    expect(legacyProbe).toHaveBeenCalledTimes(1);
-  });
-
-  it("replays through curl when DNS pre-resolution exceeds its deadline", async () => {
-    const legacyProbe: OpenAiValidationSessionDeps["legacyProbe"] = vi.fn(() => ({
-      ok: false,
-      message: "curl DNS diagnostic",
-    }));
-    const lookup = vi.fn(() => new Promise<Array<{ address: string; family: number }>>(() => {}));
-    const harness = deps(legacyProbe);
-    harness.sessionOptions = { env: {}, lookup, dnsTimeoutMs: 10 };
-
-    const result = await probeOpenAiLikeEndpointWithValidationSession(
-      "https://provider.example.test/v1",
-      "test-model",
-      "test-key",
-      {},
-      harness,
-    );
-
-    expect(result).toEqual({ ok: false, message: "curl DNS diagnostic" });
-    expect(lookup).toHaveBeenCalledTimes(1);
-    expect(legacyProbe).toHaveBeenCalledTimes(1);
-  });
-
-  it("replays through curl after a connection reset during Responses streaming", async () => {
-    const handleRequest = vi
-      .fn()
-      .mockImplementationOnce((_request, response) => {
-        response.end('{"output":[{"type":"message"}]}');
-      })
-      .mockImplementationOnce((request) => {
-        request.socket.destroy();
-      });
-    const server = http.createServer((request, response) => {
-      request.resume();
-      handleRequest(request, response);
-    });
-    const port = await listen(server);
-    const legacyProbe: OpenAiValidationSessionDeps["legacyProbe"] = vi.fn(() => ({
-      ok: false,
-      message: "curl streaming diagnostic",
-    }));
-    const harness = deps(legacyProbe);
-
-    const result = await probeOpenAiLikeEndpointWithValidationSession(
-      `http://provider.example.test:${port}/v1`,
-      "test-model",
-      "test-key",
-      { probeStreaming: true },
-      harness,
-    );
-
-    expect(result).toEqual({ ok: false, message: "curl streaming diagnostic" });
-    expect(handleRequest).toHaveBeenCalledTimes(2);
-    expect(legacyProbe).toHaveBeenCalledTimes(1);
-  });
-
-  it("uses curl without DNS pre-resolution when a proxy is configured", async () => {
-    const legacyProbe: OpenAiValidationSessionDeps["legacyProbe"] = vi.fn(() => ({
-      ok: true,
-      api: "openai-completions",
-    }));
-    const lookup = vi.fn();
-    const harness = deps(legacyProbe);
-    harness.sessionOptions = {
-      env: { HTTPS_PROXY: "http://proxy.example.test:8080" },
-      lookup,
-    };
-
-    const result = await probeOpenAiLikeEndpointWithValidationSession(
-      "https://provider.example.test/v1",
-      "test-model",
-      "test-key",
-      {},
-      harness,
-    );
-
-    expect(result).toMatchObject({ ok: true, api: "openai-completions" });
-    expect(legacyProbe).toHaveBeenCalledTimes(1);
-    expect(lookup).not.toHaveBeenCalled();
-  });
-
-  it.each([
-    "CURL_CA_BUNDLE",
-    "SSL_CERT_FILE",
-    "SSL_CERT_DIR",
-  ])("uses curl without DNS pre-resolution when %s is configured", async (envName) => {
-    const legacyProbe: OpenAiValidationSessionDeps["legacyProbe"] = vi.fn(() => ({
-      ok: true,
-      api: "openai-completions",
-    }));
-    const lookup = vi.fn();
-    const harness = deps(legacyProbe);
-    harness.sessionOptions = { env: { [envName]: "/tmp/provider-tls-config" }, lookup };
-
-    const result = await probeOpenAiLikeEndpointWithValidationSession(
-      "https://provider.example.test/v1",
-      "test-model",
-      "test-key",
-      {},
-      harness,
-    );
-
-    expect(result).toMatchObject({ ok: true, api: "openai-completions" });
-    expect(legacyProbe).toHaveBeenCalledTimes(1);
-    expect(lookup).not.toHaveBeenCalled();
-  });
-
-  it("keeps preflight-pinned endpoints on curl without native DNS", async () => {
-    const legacyProbe: OpenAiValidationSessionDeps["legacyProbe"] = vi.fn(() => ({
-      ok: true,
-      api: "openai-completions",
-    }));
-    const harness = deps(legacyProbe);
-
-    const result = await probeOpenAiLikeEndpointWithValidationSession(
-      "https://provider.example.test/v1",
-      "test-model",
-      "test-key",
-      { pinnedAddresses: ["203.0.113.10"] },
-      harness,
-    );
-
-    expect(result).toMatchObject({ ok: true, api: "openai-completions" });
-    expect(legacyProbe).toHaveBeenCalledTimes(1);
-    expect(harness.sessionOptions!.lookup).not.toHaveBeenCalled();
-  });
-
-  it("keeps DeepSeek V4 Pro on its specialized legacy streaming probe", async () => {
-    const legacyProbe: OpenAiValidationSessionDeps["legacyProbe"] = vi.fn(() => ({
-      ok: true,
-      api: "openai-completions",
-    }));
-    const harness = deps(legacyProbe);
-
-    const result = await probeOpenAiLikeEndpointWithValidationSession(
-      "https://provider.example.test/v1",
-      "deepseek-ai/deepseek-v4-pro",
-      "test-key",
-      {},
-      harness,
-    );
-
-    expect(result).toMatchObject({ ok: true, api: "openai-completions" });
-    expect(legacyProbe).toHaveBeenCalledTimes(1);
-    expect(harness.sessionOptions!.lookup).not.toHaveBeenCalled();
-  });
-
-  it("keeps query-parameter authentication out of request headers", async () => {
-    let observedUrl = "";
-    let observedAuthorization: string | undefined;
-    const server = http.createServer((request, response) => {
-      observedUrl = request.url ?? "";
-      observedAuthorization = request.headers.authorization;
-      request.resume();
-      response.end('{"choices":[{"message":{"content":"OK"}}]}');
-    });
-    const port = await listen(server);
-    const harness = deps();
-
-    const result = await probeOpenAiLikeEndpointWithValidationSession(
-      `http://provider.example.test:${port}/v1`,
-      "test-model",
-      "query-secret",
-      { authMode: "query-param", skipResponsesProbe: true },
-      harness,
-    );
-
-    expect(result).toMatchObject({ ok: true, api: "openai-completions" });
-    expect(observedAuthorization).toBeUndefined();
-    expect(new URL(observedUrl, "http://provider.example.test").searchParams.get("key")).toBe(
-      "query-secret",
-    );
-  });
-
-  it("preserves provider extra headers on the native validation request", async () => {
-    let observedReferer: string | undefined;
-    let observedTitle: string | undefined;
-    const server = http.createServer((request, response) => {
-      observedReferer = request.headers["http-referer"] as string | undefined;
-      observedTitle = request.headers["x-openrouter-title"] as string | undefined;
-      request.resume();
-      response.end('{"choices":[{"message":{"content":"OK"}}]}');
-    });
-    const port = await listen(server);
-    const harness = deps();
-
-    const result = await probeOpenAiLikeEndpointWithValidationSession(
-      `http://provider.example.test:${port}/v1`,
-      "test-model",
-      "test-key",
-      {
-        skipResponsesProbe: true,
-        extraHeaders: [
-          "HTTP-Referer: https://github.com/NVIDIA/NemoClaw",
-          "X-OpenRouter-Title: NemoClaw",
-        ],
-      },
-      harness,
-    );
-
-    expect(result).toMatchObject({ ok: true, api: "openai-completions" });
-    expect(observedReferer).toBe("https://github.com/NVIDIA/NemoClaw");
-    expect(observedTitle).toBe("NemoClaw");
     expect(harness.legacyProbe).not.toHaveBeenCalled();
   });
 });
